test: deduplicate sign helper
