+++ tests/auth.routes.test.js
@@ -2,19 +2,8 @@ import { describe, it, expect, beforeEach, afterEach, vi } from 'vitest';
 import request from 'supertest';
 import { app, User } from '../server.js';
 import bcrypt from 'bcryptjs';
-import crypto from 'crypto';
+import { sign } from './helpers/sign.js';
 
-function sign(payload, secret) {
-  const header = Buffer.from(
-    JSON.stringify({ alg: 'HS256', typ: 'JWT' }),
-  ).toString('base64url');
-  const body = Buffer.from(JSON.stringify(payload)).toString('base64url');
-  const signature = crypto
-    .createHmac('sha256', secret)
-    .update(`${header}.${body}`)
-    .digest('base64url');
-  return `${header}.${body}.${signature}`;
-}
 
 describe('auth endpoints', () => {
   let originalR2;
